# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Horse Racing AI Predictor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Track {
  id: number;
  name: string;
  location: string;
  state: string;
  surface: string;
  createdAt: string;
}

export type RaceStatus = (typeof RaceStatus)[keyof typeof RaceStatus];

export const RaceStatus = {
  upcoming: "upcoming",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Race {
  id: number;
  trackId: number;
  trackName?: string;
  raceNumber: number;
  raceName: string;
  raceDate: string;
  postTime?: string;
  distance: string;
  surface: string;
  purse?: number;
  conditions?: string;
  status: RaceStatus;
  winnerHorseId?: number | null;
  winnerHorseName?: string | null;
  createdAt: string;
}

export type CreateRaceBodySurface =
  (typeof CreateRaceBodySurface)[keyof typeof CreateRaceBodySurface];

export const CreateRaceBodySurface = {
  dirt: "dirt",
  turf: "turf",
  synthetic: "synthetic",
} as const;

export interface CreateRaceBody {
  trackId: number;
  raceNumber: number;
  raceName: string;
  raceDate: string;
  postTime?: string;
  distance: string;
  surface: CreateRaceBodySurface;
  purse?: number;
  conditions?: string;
}

export interface RecordResultBody {
  winnerHorseId: number;
  finishOrder?: number[];
}

export interface RaceEntry {
  id: number;
  raceId: number;
  horseId: number;
  horseName: string;
  postPosition: number;
  jockey: string;
  trainer: string;
  morningLineOdds?: string;
  weight?: number;
  lastRaceDate?: string | null;
  lastRaceFinish?: number | null;
  winPercentage?: number | null;
  placePercentage?: number | null;
  totalRaces: number;
  totalWins: number;
}

export interface Horse {
  id: number;
  name: string;
  age: number;
  sex: string;
  color?: string;
  sire?: string;
  dam?: string;
  trainer?: string;
  owner?: string;
  totalRaces: number;
  totalWins: number;
  totalPlaces: number;
  totalShows: number;
  earnings: number;
  createdAt: string;
}

export interface RaceHistoryEntry {
  raceId: number;
  raceName: string;
  trackName: string;
  raceDate: string;
  distance: string;
  surface: string;
  finishPosition: number;
  jockey: string;
  weight: number;
  odds?: string;
  conditions?: string;
}

export type HorseDetail = Horse & {
  recentRaces: RaceHistoryEntry[];
};

export type CreateHorseBodySex =
  (typeof CreateHorseBodySex)[keyof typeof CreateHorseBodySex];

export const CreateHorseBodySex = {
  colt: "colt",
  filly: "filly",
  gelding: "gelding",
  mare: "mare",
  horse: "horse",
} as const;

export interface CreateHorseBody {
  name: string;
  age: number;
  sex: CreateHorseBodySex;
  color?: string;
  sire?: string;
  dam?: string;
  trainer?: string;
  owner?: string;
}

export interface PredictionPick {
  horseId: number;
  horseName: string;
  rank: number;
  confidenceScore: number;
  keyFactors: string[];
}

export interface Prediction {
  id: number;
  raceId: number;
  raceName?: string;
  trackName?: string;
  raceDate?: string;
  predictedWinnerId: number;
  predictedWinnerName: string;
  confidenceScore: number;
  reasoning: string;
  topPicks: PredictionPick[];
  wasCorrect?: boolean | null;
  actualWinnerId?: number | null;
  actualWinnerName?: string | null;
  createdAt: string;
}

export interface GeneratePredictionBody {
  raceId: number;
}

export interface PredictionStats {
  totalPredictions: number;
  correctPredictions: number;
  accuracyPercentage: number;
  averageConfidence: number;
  recentAccuracy: number;
}

export type ListRacesParams = {
  trackId?: number;
  date?: string;
  status?: ListRacesStatus;
};

export type ListRacesStatus =
  (typeof ListRacesStatus)[keyof typeof ListRacesStatus];

export const ListRacesStatus = {
  upcoming: "upcoming",
  completed: "completed",
  all: "all",
} as const;

export type ListHorsesParams = {
  search?: string;
};

export type ListPredictionsParams = {
  raceId?: number;
};