# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Horse Racing AI Predictor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /horses:
    get:
      operationId: listHorses
      tags: [horses]
      summary: List horses with optional search
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of horses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Horse"
    post:
      operationId: createHorse
      tags: [horses]
      summary: Create a new horse record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHorseBody"
      responses:
        "201":
          description: Created horse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Horse"
  /horses/{id}:
    get:
      operationId: getHorse
      tags: [horses]
      summary: Get horse details and history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Horse details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HorseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /predictions:
    get:
      operationId: listPredictions
      tags: [predictions]
      summary: List all predictions
      parameters:
        - name: raceId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prediction"
  /predictions/generate:
    post:
      operationId: generatePrediction
      tags: [predictions]
      summary: Generate AI prediction for a race
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePredictionBody"
      responses:
        "200":
          description: Generated prediction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Horse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        sex:
          type: string
        color:
          type: string
        sire:
          type: string
        dam:
          type: string
        trainer:
          type: string
        owner:
          type: string
        totalRaces:
          type: integer
        totalWins:
          type: integer
        totalPlaces:
          type: integer
        totalShows:
          type: integer
        earnings:
          type: integer
        createdAt:
          type: string
      required: [id, name, age, sex, totalRaces, totalWins, totalPlaces, totalShows, earnings, createdAt]
    HorseDetail:
      allOf:
        - $ref: "#/components/schemas/Horse"
        - type: object
          properties:
            recentRaces:
              type: array
              items:
                $ref: "#/components/schemas/RaceHistoryEntry"
          required: [recentRaces]
    RaceHistoryEntry:
      type: object
      properties:
        raceId:
          type: integer
        raceName:
          type: string
        trackName:
          type: string
        raceDate:
          type: string
        distance:
          type: string
        surface:
          type: string
        finishPosition:
          type: integer
        jockey:
          type: string
        weight:
          type: integer
        odds:
          type: string
        conditions:
          type: string
      required: [raceId, raceName, trackName, raceDate, distance, surface, finishPosition, jockey, weight]
    CreateHorseBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        sex:
          type: string
          enum: [colt, filly, gelding, mare, horse]
        color:
          type: string
        sire:
          type: string
        dam:
          type: string
        trainer:
          type: string
        owner:
          type: string
      required: [name, age, sex]
    Prediction:
      type: object
      properties:
        id:
          type: integer
        raceId:
          type: integer
        raceName:
          type: string
        trackName:
          type: string
        raceDate:
          type: string
        predictedWinnerId:
          type: integer
        predictedWinnerName:
          type: string
        confidenceScore:
          type: number
        reasoning:
          type: string
        topPicks:
          type: array
          items:
            $ref: "#/components/schemas/PredictionPick"
        wasCorrect:
          type: boolean
          nullable: true
        actualWinnerId:
          type: integer
          nullable: true
        actualWinnerName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, raceId, predictedWinnerId, predictedWinnerName, confidenceScore, reasoning, topPicks, createdAt]
    PredictionPick:
      type: object
      properties:
        horseId:
          type: integer
        horseName:
          type: string
        rank:
          type: integer
        confidenceScore:
          type: number
        keyFactors:
          type: array
          items:
            type: string
      required: [horseId, horseName, rank, confidenceScore, keyFactors]
    GeneratePredictionBody:
      type: object
      properties:
        raceId:
          type: integer
      required: [raceId]
...