# lib/api-zod/src/generated/types/createHorseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Horse Racing AI Predictor API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateHorseBodySex } from "./createHorseBodySex";

export interface CreateHorseBody {
  name: string;
  age: number;
  sex: CreateHorseBodySex;
  color?: string;
  sire?: string;
  dam?: string;
  trainer?: string;
  owner?: string;
}