# lib/api-zod/src/generated/types/horse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Horse Racing AI Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface Horse {
  id: number;
  name: string;
  age: number;
  sex: string;
  color?: string;
  sire?: string;
  dam?: string;
  trainer?: string;
  owner?: string;
  totalRaces: number;
  totalWins: number;
  totalPlaces: number;
  totalShows: number;
  earnings: number;
  createdAt: string;
}