# lib/api-zod/src/generated/types/horseDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Horse Racing AI Predictor API
 * OpenAPI spec version: 0.1.0
 */
import type { Horse } from "./horse";
import type { RaceHistoryEntry } from "./raceHistoryEntry";

export type HorseDetail = Horse & {
  recentRaces: RaceHistoryEntry[];
};