# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Horse Racing AI Predictor API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionPick } from "./predictionPick";

export interface Prediction {
  id: number;
  raceId: number;
  raceName?: string;
  trackName?: string;
  raceDate?: string;
  predictedWinnerId: number;
  predictedWinnerName: string;
  confidenceScore: number;
  reasoning: string;
  topPicks: PredictionPick[];
  wasCorrect?: boolean | null;
  actualWinnerId?: number | null;
  actualWinnerName?: string | null;
  createdAt: string;
}