# lib/api-zod/src/generated/types/predictionPick.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Horse Racing AI Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionPick {
  horseId: number;
  horseName: string;
  rank: number;
  confidenceScore: number;
  keyFactors: string[];
}